+++ Src/Private/Get-AbrWinFOClusterQuorum.ps1
@@ -5,7 +5,7 @@ function Get-AbrWinFOClusterQuorum {
     .DESCRIPTION
         Documents the configuration of Microsoft Windows Server in Word/HTML/Text formats using PScribo.
     .NOTES
-        Version:        0.5.0
+        Version:        0.5.2
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -21,16 +21,16 @@ function Get-AbrWinFOClusterQuorum {
 
     begin {
         Write-PScriboMessage "FailOverCluster InfoLevel set at $($InfoLevel.FailOverCluster)."
-        Write-PscriboMessage "Collecting Host FailOver Cluster Quorum information."
+        Write-PScriboMessage "Collecting Host FailOver Cluster Quorum information."
     }
 
     process {
         try {
-            $Settings = Invoke-Command -Session $TempPssSession { Get-ClusterQuorum -Cluster $using:Cluster | Select-Object -Property * } | Sort-Object -Property Name
+            $Settings = Invoke-Command -Session $TempPssSession { Get-ClusterQuorum | Select-Object -Property * } | Sort-Object -Property Name
             if ($Settings) {
                 Section -Style Heading3 "Quorum" {
                     $OutObj = @()
-                    foreach  ($Setting in $Settings) {
+                    foreach ($Setting in $Settings) {
                         try {
                             $inObj = [ordered] @{
                                 'Cluster' = $Setting.Cluster
@@ -38,14 +38,13 @@ function Get-AbrWinFOClusterQuorum {
                                 'Quorum Type' = $Setting.QuorumType
                             }
                             $OutObj += [pscustomobject]$inobj
-                        }
-                        catch {
-                            Write-PscriboMessage -IsWarning $_.Exception.Message
+                        } catch {
+                            Write-PScriboMessage -IsWarning $_.Exception.Message
                         }
                     }
 
                     $TableParams = @{
-                        Name = "Quorum - $($Cluster.toUpper().split(".")[0])"
+                        Name = "Quorum - $($Cluster)"
                         List = $false
                         ColumnWidths = 33, 34, 33
                     }
@@ -55,9 +54,8 @@ function Get-AbrWinFOClusterQuorum {
                     $OutObj | Table @TableParams
                 }
             }
-        }
-        catch {
-            Write-PscriboMessage -IsWarning $_.Exception.Message
+        } catch {
+            Write-PScriboMessage -IsWarning $_.Exception.Message
         }
     }