+++ Src/Private/Get-AbrWinNetFirewall.ps1
@@ -5,7 +5,7 @@ function Get-AbrWinNetFirewall {
     .DESCRIPTION
         Documents the configuration of Microsoft Windows Server in Word/HTML/Text formats using PScribo.
     .NOTES
-        Version:        0.5.0
+        Version:        0.5.2
         Author:         Andrew Ramsay
         Editor:         Jonathan Colon
         Twitter:        @asbuiltreport