+++ Src/Private/Get-AbrWinOSConfig.ps1
@@ -5,7 +5,7 @@ function Get-AbrWinOSConfig {
     .DESCRIPTION
         Documents the configuration of Microsoft Windows Server in Word/HTML/Text formats using PScribo.
     .NOTES
-        Version:        0.2.0
+        Version:        0.5.2
         Author:         Andrew Ramsay
         Editor:         Jonathan Colon
         Twitter:        @asbuiltreport
@@ -21,27 +21,35 @@ function Get-AbrWinOSConfig {
 
     begin {
         Write-PScriboMessage "Operating System InfoLevel set at $($InfoLevel.OperatingSystem)."
-        Write-PscriboMessage "Collecting Oprating System Configuration information."
+        Write-PScriboMessage "Collecting Oprating System Configuration information."
     }
 
     process {
         if ($InfoLevel.OperatingSystem -ge 1) {
             Section -Style Heading3 'OS Configuration' {
                 Paragraph 'The following section details hos OS configuration'
-                Blankline
+                BlankLine
                 $HostOSReport = [PSCustomObject] @{
-                'Windows Product Name' = $HostInfo.WindowsProductName
-                'Windows Version' = $HostInfo.WindowsCurrentVersion
-                'Windows Build Number' = $HostInfo.OsVersion
-                'Windows Install Type' = $HostInfo.WindowsInstallationType
-                'AD Domain' = $HostInfo.CsDomain
-                'Windows Installation Date' = switch (($HostInfo.OsInstallDate).count) {
-                    0 {'-'}
-                    default {$HostInfo.OsInstallDate.ToShortDateString()}
-                }
-                'Time Zone' = $HostInfo.TimeZone
-                'License Type' = $HostLicense.ProductKeyChannel
-                'Partial Product Key' = $HostLicense.PartialProductKey
+                    'Windows Product Name' = $HostInfo.WindowsProductName
+                    'Windows Version' = $HostInfo.WindowsCurrentVersion
+                    'Windows Build Number' = $HostInfo.OsVersion
+                    'Windows Install Type' = $HostInfo.WindowsInstallationType
+                    'AD Domain' = $HostInfo.CsDomain
+                    'Windows Installation Date' = switch (($HostInfo.OsInstallDate).count) {
+                        0 { "--" }
+                        default { $HostInfo.OsInstallDate.ToShortDateString() }
+                    }
+                    'Time Zone' = $HostInfo.TimeZone
+                    'License Type' = Switch ([string]::IsNullOrEmpty($HostLicense.ProductKeyChannel)) {
+                        $true { "--" }
+                        $false { $HostLicense.ProductKeyChannel }
+                        default { "Unknown" }
+                    }
+                    'Partial Product Key' = Switch ([string]::IsNullOrEmpty($HostLicense.PartialProductKey)) {
+                        $true { "--" }
+                        $false { $HostLicense.PartialProductKey }
+                        default { "Unknown" }
+                    }
                 }
                 $TableParams = @{
                     Name = "OS Settings"