+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # :arrows_counterclockwise: Microsoft Windows As Built Report Changelog
 
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.5.2] - 2024-02-18
+
+### Added
+
+- Add Local Windows Group Members information @flynngw
+- Add more HealthCheck conditions
+
+### Changed
+
+- Improve report readability
+
+### Fixed
+
+- Fix CodeQL security alerts
+- Fix FailOver Cluster section
+- Fix issue in Get-AbrWinOSHotfix install date logic
+
 ## [0.5.1] - 2023-05-16
 
 ### Added